Scaffolding: Add new `keywords`section, and placeholder screenshot.
This introduces the concept of adding keywords and proper descriptions when creating blocks, allowing for a more automated means of discovering what other code we have previously written across our projects.

By including a screenshot as part of that expectation, we can surface block solutions and their visuals in a more coherent way that is easier to understand and visualize regardless of technical competency.